+++ client/gesture/gesture.dart
@@ -12,8 +12,11 @@
 #import("../html/html.dart");
 #import("../effect/effect.dart");
 
+#source("src/Gesture.dart");
 #source("src/HoldGesture.dart");
 #source("src/DragGesture.dart");
 #source("src/SwipeGesture.dart");
 #source("src/ZoomGesture.dart");
+
+#source("src/Dragger.dart");
 #source("src/Scroller.dart");